feat: Add first support of negative offsets
There are two ways to make pointer treated as pointer to field in
the structure:
1) If function has CONTAINING_RECORD macro recognized by IDA
(his can be manually made by changing types), then pointer
will be converted to CONTAINING_RECORD macro wherever it
needed in the text.
2) If local variable has comment ```Structure+X```, where
Structure is type of containing structure and X is offset to the field
in bytes.